recommendation-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
faiss-cpu>=1.13.2
# sentence-transformers 2.2.x требует huggingface_hub с cached_download (удалён в 0.20+)
huggingface_hub>=0.17.0,<0.20.0
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
httpx==0.26.0
prometheus-client==0.21.0
